Further made the minimised scaling transition weaker
There's positioning issues with clicking the checkbox on an entry (clicking on it cancels out when the mouse goes out of its bounds, and it happens when the entry itself scales a bit away from 100%). This even more reduced motion should reduce the chances of the button cancelling (unless the window/screen is too wide).